+++ src/ophyd_async/epics/demo/_ioc.py
@@ -1,6 +1,8 @@
+import subprocess
+import sys
 from pathlib import Path
 
-from ophyd_async.epics.testing import Database
+from ophyd_async.epics.testing import DEFAULT_SOFTIOC_ARGS, Database, ioc_argv
 
 HERE = Path(__file__).absolute().parent
 
@@ -35,3 +37,23 @@ def demo_ioc_database(prefix: str, num_channels: int) -> list[Database]:
         for i in range(1, num_channels + 1)
     ]
     return databases
+
+
+if __name__ == "__main__":
+    # Convenience standalone entry point: `python _ioc.py <prefix> <num_channels>
+    # [softioc_args...]` builds and hosts the demo topology directly, without
+    # needing to write any Python. Trailing args override which executable
+    # actually hosts the IOC, defaulting to DEFAULT_SOFTIOC_ARGS - e.g.
+    # `python _ioc.py demo: 3 softIoc` to use a real EPICS installation instead
+    # of the bundled epicscorelibs.ioc.
+    if len(sys.argv) < 3:
+        raise SystemExit(
+            f"Usage: {sys.argv[0]} <prefix> <num_channels> [softioc_args...]"
+        )
+    _prefix, _num_channels, *_softioc_args = sys.argv[1:]
+    _databases = demo_ioc_database(_prefix, int(_num_channels))
+    sys.exit(
+        subprocess.run(
+            [*(_softioc_args or DEFAULT_SOFTIOC_ARGS), *ioc_argv(_databases)]
+        ).returncode
+    )
