+++ src/ophyd_async/epics/testing/_utils.py
@@ -15,6 +15,11 @@ DEFAULT_SOFTIOC_ARGS: Sequence[str] = (sys.executable, "-m", "epicscorelibs.ioc"
 _READY_MARKER = "iocRun: All initialization complete"
 _STOP_INPUT = "exit()"
 
+# _ioc.py forwards argv verbatim to whatever hosts the IOC, so every start_ioc
+# call is a real invocation of this file *by path* - the same command a human
+# would type (see _ioc.py's module docstring), not something built up inline.
+_IOC_PY = str(Path(__file__).parent / "_ioc.py")
+
 
 def generate_random_pv_prefix() -> str:
     """For generating random PV names in test devices."""
@@ -27,6 +32,19 @@ class Database:
 
     path: Path | str
     macros: dict[str, str]
+
+
+def ioc_argv(databases: Sequence[Database]) -> list[str]:
+    """Flatten `databases` into the `-m macro -d db.db [...]` argv.
+
+    The shape softIoc/`epicscorelibs.ioc` expect - the part of the command
+    line after `--` in `_ioc.py`'s own usage.
+    """
+    argv: list[str] = []
+    for db in databases:
+        macro_str = ",".join(f"{k}={v}" for k, v in db.macros.items())
+        argv += ["-m", macro_str, "-d", str(db.path)]
+    return argv
 
 
 def start_ioc(
@@ -45,8 +63,11 @@ def start_ioc(
         bundled `epicscorelibs.ioc`. Override to run against a real EPICS
         installation's `softIoc` binary instead, e.g. `["softIoc"]`.
     """
-    args = list(softioc_args)
-    for db in databases:
-        macro_str = ",".join(f"{k}={v}" for k, v in db.macros.items())
-        args += ["-m", macro_str, "-d", str(db.path)]
-    return start_subprocess(args, _READY_MARKER, stop_input=_STOP_INPUT)
+    subprocess_args = [
+        sys.executable,
+        _IOC_PY,
+        *softioc_args,
+        "--",
+        *ioc_argv(databases),
+    ]
+    return start_subprocess(subprocess_args, _READY_MARKER, stop_input=_STOP_INPUT)
