+++ src/ophyd_async/tango/testing/__init__.py
@@ -1,13 +1,10 @@
 from ._example_types import ExampleStrEnum, generate_random_trl_prefix
-from ._tango_device_servers import (
-    DEFAULT_PYTHON_ARGS,
-    predict_trl,
-    start_tango_device_servers,
-    tango_device_servers_args,
-)
+from ._launch import DEFAULT_PYTHON_ARGS, Catalog, start_tango_device_servers
+from ._tango_device_servers import predict_trl, tango_device_servers_args
 from ._test_device import TangoTestDevice
 
 __all__ = [
+    "Catalog",
     "DEFAULT_PYTHON_ARGS",
     "ExampleStrEnum",
     "TangoTestDevice",
